Closer to JTS 1.3
git-svn-id: http://svn.osgeo.org/geos/trunk@120 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/graph/EdgeList.cpp
@@ -15,7 +15,7 @@ void EdgeList::insert(Edge *e) {
  *          -1 otherwise
  */
 int EdgeList::findEdgeIndex(Edge *e) {
-	for (int i=0; i<this->size();i++) {
+	for (int i=0; i<(int)this->size();i++) {
 		if ( (*this)[i]->equals(e) )
             return i;
 	}
